move _as_qubit_iterable to types/qubits.py
